修复点击表单出现404页面的错误
