+++ 23-26-real-project/src/containers/index.js
@@ -2,6 +2,6 @@ export Main from './Main/Main';
 export Home from './Home/Home';
 export NotFound from './NotFound/NotFound';
 export Counter from './Counter/Counter';
-export Forms from './Forms/Form';
+export Forms from './Forms/Forms';
 export Statistic from './Statistic/Statistic';
 export Login from './Login/Login';
