+++ 23-26-real-project/webpack-assets.json
@@ -1,6 +1,6 @@
 {
   "javascript": {
-    "main": "http://localhost:3003/dist/main-207f4388eef1c170f867.js"
+    "main": "http://localhost:3003/dist/main-0e324d2740ea475e8195.js"
   },
   "styles": {},
   "assets": {
